Refactorizo el codigo un poco para que este mas escalable con las capas internas

+++ src/main/java/com/auth/wow/libre/infrastructure/repositories/account_web/AccountWebRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AccountWebRepository extends CrudRepository<AccountWebEntity, Integer> {
+public interface AccountWebRepository extends CrudRepository<AccountWebEntity, Long> {
+
 }
